+++ src/com/learn/java/day4/Biggest.java
@@ -4,8 +4,8 @@ import java.util.*;
 public class Biggest {
 	public static void main(String[] args) {
 		Scanner sc=new Scanner(System.in);
-		int a=sc.nextInt();
-		int b=sc.nextInt();
+		double a=sc.nextDouble();
+		double b=sc.nextDouble();
 		System.out.println("The biggest number is: "+((a>b)?a:b));
 	}
