+++ src/com/learn/java/day4/LargestAndSmallest.java
@@ -5,9 +5,9 @@ import java.util.Scanner;
 public class LargestAndSmallest {
 	public static void main(String[] args) {
 		Scanner sc=new Scanner(System.in);
-		int a=sc.nextInt();
-		int b=sc.nextInt();
-		int c=sc.nextInt();
+		double a=sc.nextDouble();
+		double b=sc.nextDouble();
+		double c=sc.nextDouble();
 		if(a>b && a>c) {
 			System.out.println("Largestnumber is "+a);
 			System.out.println("Smallest number is "+((b>c)?c:b));
