+++ src/com/learn/java/day7/TotalAndAverageOfNumbers.java
@@ -6,11 +6,11 @@ public class TotalAndAverageOfNumbers {
 	public static void main(String[] args) {
 		Scanner sc=new Scanner(System.in);
 		System.out.print("Enter n: ");
-		float n=sc.nextInt();
-		int sum=0;
+		int n=sc.nextInt();
+		double sum=0;
 		System.out.print("Enter the numbers: ");
 		for(int i=1;i<=n;i++) {
-			sum+=sc.nextInt();
+			sum+=sc.nextDouble();
 		}
 		System.out.print("sum of the numbers is: "+sum+"\n"+"Average= "+(sum/n));
 	}
